bf2mods: Fix moonjump cam, clean up movement hooks & keybind code
Finally fixes the player camera when moonjumping. To do this, we force the player camera's target to always pretend like it is on the ground, and that the height of the surface the player should be walking on is set to the height of the control mover.

Movement code has also been moved out of debug_stuff and put into its own file, player_movement. Keybind handling was also touched up a bit, though I may end up also separating it out into its own file.