*: Working CMake "port"
This was originally created for "better" CLion tooling, but after multiple hours (over nearly a year) of bashing our collective heads
with keyboards, we finally got it building properly in CMake (and it was one silly issue in hindsight).

The residual Makefile is solely for automating ftp transfer (and probably doesn't even work atm), even then it might be removed later.