bf2mods: Working engine-level freecam, remove event freecam
For this, some classes were filled out and input was partially redone. All binds and button functions were moved to plugin_main's header for ease of access in other namespaces.

Most of the mm:: hand-rolled math types were replaced with typedefs to the GLM library, for ease of use in math operations.

Calls to Skyline's logger in hook setups were also replaced with the custom logger, as they appear during boot, as well as trickle down to the Skyline logger.

send_over_ftp.py was also slightly edited to remove the 3rd argument as the built executable has the same name no matter the code name.