Fix freecam for DE, implement codename defines, update README
DE's engine changed quite a bit internally compared to 2 or Torna. Framework's update functionality was split out into a seprate FrameworkUpdater class, and I'm unsure if that's because of legacy 1 code or an engine-level refactor. Either way, a different update function needs to be hooked for DE.

DE's camera also uses ScnObjCam's setWorldMatrix, as opposed to setViewMatrix like in 2 or Torna. This means that we don't have to inverse the matrix to get world space out, as we're in it already.

Also might as well add a dedicated Torna codename (ira) npdm.