bf2mods: Remove plugin class, SharedState; fix emulators and in-game version
SharedState has been unshared, it's now just an allocated struct. Previously the call to shmemCreate would crash emulators (maybe some real hardware too?) due to a missing service call, so we wanted to get that removed.

In doing so, we realized plugin.cpp no longer had a purpose, so it was removed as well.

The in-game version was broken as well, it now just uses a small char buffer to hold the git version. Might as well print that information to the log too, right?