bf2mods: Fix bf2logger bugs, add _DEBUG define
bf2logger suffered from a bug where some messages would get eaten, with seemingly no cause. The code was such that when the list overflowed and removed the oldest message, it would neglect to add the new one it was supposed to. Another smaller bug was fixed relating to drawing and underflowing lifetimes.

A _DEBUG define was also added for debug builds. NDEBUG is its release equivalent.

Co-authored-by: modeco80 <lily.modeco80@protonmail.ch>