bf2mods: Allow 2+ controllers in game, hide debug draws with UI
The P2 controller for keybinds can now be connected while the game is running, as previously the OS was told the game was for a single controller only.

Additionally, the pad debug draw and BgmTrack debug draw are now turned off when the UI is also off.

Also a small line shift accident caused player_movement to be compiled for DE, whoops.