bf2mods: Math types inherit from glm, always use world-space for freecam
mm's math types can now be used properly in function prototypes we have rtld fill out. Before, the compiler/linker would generate imports for their glm types, instead of the desired mm type. Some static pointer juggling is needed in some instances, but the tradeoff is worth it.

The DoFreeCameraMovement function has been updated to always be in world-space coordinates, as the hook function now does the "heavy lifting" of matrix inversion. This was really only done to keep the code neater, and allow the hook function to draw cameras.

Co-authored-by: modeco80 <lily.modeco80@protonmail.ch>