bf2mods: Preliminary file read interception (not like LayeredFS)
Using the nicety of having an SD card, we intercept calls that use ml::DevFileTh::readFile and dump the read (possibly decompressed in some cases) data into a folder.

However, not all files are read through this method, and even still some reads are not complete by the time we try to dump the data. More research was done and we found the actual threaded read function, that'll come next commit.

Co-authored-by: modeco80 <lily.modeco80@protonmail.ch>