bf2mods: Config for event debug displays, use the port for TCP
Slight refactor here for the skylaunch logger, it now can take a port in its constructor, which it fetches from the config. The ctor and SD mount were moved to the ROM mount hook we don't have to worry about the SD not being present.

Also fixed a tiny bug so index 0 can actually be toggled.