bf2mods: Better titlescreen version display
I was looking at Mod-Ardain by @RoccoDev, and noticed their version display actually creates a whole new UI element! It's much nicer than having to hack up a .wilay or .uio so I've adopted my own version here.

DE looks like its UI system works differently, so it may take a bit to make a version of this for that.

The version in BdatRandomizer was removed in favor of this.