bf2mods: Don't allow titlescreen event changing on new game
Went to go try a new game and it loaded a random chapter cutscene then left me at the origin of that map. Apparently that uses the same method to load the event, so let's maybe not do that in the future.

For the hacky way I do this I needed to add an overload to dbgutil::getSymbol to retain mangled names.