bf2mods: More UI object research, build date on titlescreen with debug
This whole UI system seems haunted in the render stage. The amount of characters that are actually rendered seems to be determined on some sort of size, but I can't find where that size is located. For now we'll just use separate objects for the build date

Fixed some memory issues with FixStr as well.