bf2mods: Add new BdatOverride module
This module is a replacement for the old MSText-only hook, providing nicer control over all returned Bdat values. It uses a simple system to invoke callbacks during value fetches and overwrite returned values or swap the row index entirely.

BdatRandomizer was switched to use this system. There's not a good way to get the original filename for a sheet, so it checks for "name" and "style" members.

Co-authored-by: modeco80 <lily.modeco80@protonmail.ch>