xm: Freecam for menu model views (models in UI)
DE uses these for the party display/setup, Event Theater background, and oddly the change time clock.

This needed some changes to the rest of the freecam code, like making DoFreeCameraMovement take in a matrix instead of always using the one in the struct, and now the freecam state has another matrix specifically for these UI elements.