xm: Logger toast messages
The little thing I had written for the freecam speed was quite nice, if a bit hobbled together. I realized how handy a system like that would be for other parts of the code, so I bolted that functionality onto the logger.

Essentially, logger messages now have a "group," used to keep track of what type of message this is. When another message is added with the same group, the first instance (if it exists) is instead updated. This isn't used for normal log messages, not unless I want to go full VConsole on this.

The freecam speed display now uses this, as well as the version printout.

Also, there's a new wrapper for drawFontGetWidth on fmt strings.