xm: Freecam for more camera types (UI models for 2)
Someone pointed out I forgot to handle the UI models for 2, so I went in and researched that. In the process I realized it would be much easier to just go even lower level and move the fw::Cameras around.

This sorta does that, just with fw::Camera::setLookAt instead of ::update due to overridden update functions.