xm: Massively simplify freecam code, truncate logMemory output
Instead of cobbling together a lot of different methods to update cameras, we now go for modifying the camera layers instead. This cleans up the code a tremendous amount, and simplifies all the actual camera puppeting down into one hook. Aside from structs and parameters, the code is even identical between 2/Torna and DE.

The code that puts the state of the current camera into the freecam has its own hook now, and I would like to do a bit more research and see what I can do to *always* make that take the active camera, but for now it just takes the main scene camera as before.

In the process of debugging this, logMemory was also updated so its output wouldn't get lobbed off due to the max log length.