xm: Another attempt at Bdat override optimization
Bdat accesses with the hook seem to be incredibly performance demanding. The biggest culprit for this is CHR_Ir in Torna, which takes up a whopping 50% of all accesses while in game. FLD_QuestListAchievement isn't too far behind in 2.

This attempts to clean that up by skipping the edits on a few specified tables. Not an ideal solution by any means, but this is behind a config option, so it can be disabled if needed.

string_views were also converted back to char*s, because there wasn't really any reason not to.