xm: 60FPS patch for DE
This behavior is already done by a few cheats, but this method works under any version xenomods supports.

Ideally this would work for 2 and Torna as well, and it *technically* does! But they do not implement frame skipping, so when the console dips below 60 (as is most often the case) the whole engine slows.

The only thing this technically breaks are events and UI, but it's better to have a proper fix than do some jank stuff to get it to work.