xm: Rewrite freecam input, account for delta time
There's some plans for freecam-adjacent nonsense, and I realized there was no centralized keybind for freecam controls. So, there's that now.

When testing the 60FPS stuff for DE, I remembered the camera was not tied to the delta time at all, which meant it would be 2x faster when at max framerate. Multiplying the speeds by the delta time fixes this.

Ironically enough, for DE, I couldn't get the delta time with what I looked at. So it's set to a fixed 1/30. Remind me to fix this.