xm: Simplify args for DoFreeCameraMovement, use delta time
It didn't need to be a member function, as it would never be used by other code. It is now only declared in the .cpp.

In the process of doing that, the arguments were cleaned up and now only an UpdateInfo is passed in, which means DE now has a framerate independent camera.