xm: Debug rendering in DE (thanks @RoccoDev!)
As usual, I've taken a thing someone else gone and did and simplified it.

Rocco figured out the method of enabling debug rendering in DE by adding a render callback and setting it to be the debdraw target. I figured out that the IScnRender that was needed was not the same as the FontLayer, but rather could be gotten from the scene's ScnDebug.