xm: Starter implementation of a menu system
We're getting to the point where there are far too many keybinds to keep track of. Even for me, I forget the ones I haven't used in a while. So, let's make a menu!

This is just a little text interface that will let the user control state and modules, along with any other things I may add in the future.

The menu will draw instead of the logger messages when it is open.

Co-authored-by: modeco80 <lily.modeco80@protonmail.ch>