xm: Implement framework for menu Options (and float editor)
This is a solid implementation, but I'm out of brain power to write a detailed commit message.

Options are refusing to allocate right now and we can't figure out why, so the examples are statically created atm. Hopefully next commit fixes that.

Also, the menu will use P1 if there's no P2 connected.

Co-authored-by: modeco80 <lily.modeco80@protonmail.ch>