xm: Rework warps, make them serializable/sharable files
Only being able to save one warp at a time and being unable to export them severely limited the actual use of warps. I thought of them as useful for trying out clips, but what's the point if they vanish after closing the game?

Now, they can be saved to a toml file. Share the file, or just copy a specific warp, and then others can use it. Voilà!