xm: IsModuleRegistered to check if a module exists, use for external module calls
As can be expected, external module calls are unsafe just due to the nature of a module system. If they aren't loaded, then the code won't be present, and rtld will hang the system. So, this helper checks the symbols that are generated for one of that module name, then returns true if it is safe to call.