xm: Move hooks in main to their own class, HidInput::GetDebugInput
main was beginning to get cluttered from all the hooks I wanted to use, so they've been moved into their own class. This caused a small issue with anonymous namespaces, but it's been solved now.

The debug input function was also something I wanted to do for a bit, it takes the last connected controller and grabs its inputs.