xm: General clean up of Menu options, update docs
This is a broad cleanup/reordering of the modules in the Menu. This comes with changing compile order, which may end up causing some issues down the road. (hopefully not!)

FileDetours' options were removed as they're already in the config.
RenderingControls had all of the skip controls moved into the Toggles... section.
PartyEditor got some renames of some options to clear up any confusion.

Docs were also updated to reflect changes, and to add the updated party editor controls.