xm: Remove Warp name editing/add delete button, clarify config errors
In the Warps dialog, the name text field and tab would be highlighted at the same time, which caused weirdness for selections. This is likely something that can be fixed, but considering the keyboard can't be used on console, it might be fine.

Config errors were also clarified as I had a few people get confused at the "Couldn't load config!" message on first boot.