skylaunch: Skyline plugin support (+ BREAKING CHANGES!)
Adds (more like backports) the functionality in mainline Skyline to allow plugins to function.

In response, subsdk1 has become subsdk9 to match other Skyline plugins. This rename is to avoid problems with drag and drops. It's better for a user to wonder why the mods don't load rather than wonder why their game is crashing on boot. Be careful to remove the existing subsdk1.